maps fragment link added to nav drawer
