+++ app/src/main/AndroidManifest.xml
@@ -11,12 +11,6 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Factotum"
         tools:targetApi="31">
-        <activity
-            android:name=".maps.BasicActivity"
-            android:exported="false"
-            android:label="@string/title_activity_basic"
-            android:theme="@style/Theme.Factotum.NoActionBar" />
-
         <meta-data
             android:name="com.google.android.gms.version"
             android:value="@integer/google_play_services_version" />
