+++ app/src/main/java/com/github/factotum_sdp/factotum/ui/maps/MapsFragment.kt
@@ -4,34 +4,53 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProvider
-import com.github.factotum_sdp.factotum.databinding.FragmentDirectoryBinding
+import androidx.fragment.app.activityViewModels
+import com.github.factotum_sdp.factotum.databinding.FragmentMapsBinding
+import com.google.android.gms.maps.CameraUpdateFactory
+import com.google.android.gms.maps.SupportMapFragment
+import com.google.android.gms.maps.model.LatLng
 
-class MapsFragment: Fragment() {
-    private var _binding: FragmentDirectoryBinding? = null
+
+/**
+ * A simple [Fragment] subclass as the second destination in the navigation.
+ */
+class MapsFragment : Fragment() {
+
+    private var _binding: FragmentMapsBinding? = null
+    private val viewModel: MapsViewModel by activityViewModels()
+
 
     // This property is only valid between onCreateView and
     // onDestroyView.
     private val binding get() = _binding!!
 
     override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
+        inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View {
-        val slideshowViewModel =
-            ViewModelProvider(this).get(MapsViewModel::class.java)
+    ): View? {
 
-        _binding = FragmentDirectoryBinding.inflate(inflater, container, false)
-        val root: View = binding.root
+        _binding = FragmentMapsBinding.inflate(inflater, container, false)
+        return binding.root
+    }
 
-        val textView: TextView = binding.textSlideshow
-        slideshowViewModel.text.observe(viewLifecycleOwner) {
-            textView.text = it
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        val mapFragment = binding.map.getFragment() as SupportMapFragment
+        mapFragment.getMapAsync{ googleMap ->
+            googleMap.clear()
+            for (route in viewModel.routes.value.orEmpty()){
+                route.addToMap(googleMap)
+            }
+            val epfl = LatLng(46.520536, 6.568318)
+            googleMap.moveCamera(CameraUpdateFactory.newLatLng(epfl))
+
+            // Add zoom controls to the map
+            googleMap.uiSettings.isZoomControlsEnabled = true
+
+            // Add zoom gestures to the map
+            googleMap.uiSettings.isZoomGesturesEnabled = true
         }
-        return root
     }
 
     override fun onDestroyView() {
