+++ app/src/main/java/com/github/factotum_sdp/factotum/ui/maps/MapsViewModel.kt
@@ -1,13 +1,18 @@
 package com.github.factotum_sdp.factotum.ui.maps
 
-import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.github.factotum_sdp.factotum.data.Route
 
 class MapsViewModel : ViewModel() {
 
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is the maps Fragment"
+    val routes: MutableLiveData<MutableList<Route>> = MutableLiveData(mutableListOf())
+
+    fun addRoute(route: Route){
+        routes.value?.add(route)
     }
-    val text: LiveData<String> = _text
+
+    fun deleteAll(){
+        routes.value = mutableListOf()
+    }
 }
