+++ app/src/main/java/com/github/factotum_sdp/factotum/ui/maps/RouteFragment.kt
@@ -1,4 +1,4 @@
-package com.github.factotum_sdp.factotum.maps
+package com.github.factotum_sdp.factotum.ui.maps
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -9,8 +9,9 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.findNavController
 import com.github.factotum_sdp.factotum.R
-import com.github.factotum_sdp.factotum.databinding.FragmentFirstBinding
-import com.google.android.gms.maps.model.LatLng
+import com.github.factotum_sdp.factotum.data.Route
+import com.github.factotum_sdp.factotum.databinding.FragmentRoutesBinding
+import com.github.factotum_sdp.factotum.ui.maps.MapsViewModel
 
 
 /**
@@ -18,8 +19,8 @@ import com.google.android.gms.maps.model.LatLng
  */
 class RouteFragment : Fragment() {
 
-    private var _binding: FragmentFirstBinding? = null
-    private val viewModel: MainView by activityViewModels()
+    private var _binding: FragmentRoutesBinding? = null
+    private val viewModel: MapsViewModel by activityViewModels()
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -30,7 +31,7 @@ class RouteFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
 
-        _binding = FragmentFirstBinding.inflate(inflater, container, false)
+        _binding = FragmentRoutesBinding.inflate(inflater, container, false)
         return binding.root
 
     }
@@ -44,7 +45,8 @@ class RouteFragment : Fragment() {
 
         viewModel.deleteAll()
         val listCourse = arrayListOf("BC -> Satelitte", "EPFL -> Chauderon", "Genève -> Lausanne")
-        val listRoute = arrayListOf<Route>(Route(46.51869448523383, 6.561842896370142, 46.520742473314236, 6.567824983999257),
+        val listRoute = arrayListOf<Route>(
+            Route(46.51869448523383, 6.561842896370142, 46.520742473314236, 6.567824983999257),
             Route(46.51916261132295, 6.566773558010879, 46.52369559941859, 6.6250423828801654),
             Route(46.205062260846894, 6.1430383670835464, 46.517234720289416, 6.6291717405531605),
         )
