+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -17,7 +17,7 @@
             android:icon="@drawable/ic_menu_camera"
             android:title="@string/menu_picture" />
         <item
-            android:id="@+id/mapsFragment"
+            android:id="@+id/routeFragment"
             android:icon="@android:drawable/ic_dialog_map"
             android:title="@string/menu_maps" />
     </group>
